ml_models/requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
matplotlib==3.8.2
seaborn==0.13.0
shap==0.44.0
